fix: 添加表格点击事件
