+++ scaleph-ui-react/src/pages/Project/Workspace/Artifact/Sql/CodeEditor/EditorRightResult/EditorRightResultTable.tsx
@@ -1,6 +1,6 @@
-import { Table } from 'antd';
+import { Button, message, Modal, Table } from 'antd';
 import type { ColumnsType } from 'antd/es/table';
-import React from 'react';
+import React, { useState } from 'react';
 import styles from './index.less';
 
 interface DataType {
@@ -10,70 +10,10 @@ interface DataType {
   address: string;
 }
 
-const columns: ColumnsType<DataType> = [
-  {
-    title: 'Name',
-    dataIndex: 'name',
-    width: 130,
-    render: (value: any, row: any, rowIndex: number) => {
-      return (
-        <div className={styles.tableItem}>
-          <div>{value}</div>
-          <div className={styles.tableHoverBox}>
-            <img
-              src="https://s.xinc818.com/files/webcilklz16y4pxm3zv/位图 (1).svg"
-              alt="查看"
-              onClick={() => {
-                console.log('查看');
-              }}
-            />
-            <img
-              src="https://s.xinc818.com/files/webcilklz19gz7rldus/复制_o.svg"
-              alt="复制"
-              onClick={() => {
-                console.log('复制');
-              }}
-            />
-          </div>
-        </div>
-      );
-    },
-    sorter: (a, b) => a.name.length - b.name.length,
-  },
-  {
-    title: 'Age',
-    dataIndex: 'age',
-    width: 130,   
-    render: (value: any, row: any, rowIndex: number) => {
-        return (
-          <div className={styles.tableItem}>
-            <div>{value}</div>
-            <div className={styles.tableHoverBox}>
-              <img
-                src="https://s.xinc818.com/files/webcilklz16y4pxm3zv/位图 (1).svg"
-                alt="查看"
-                onClick={() => {
-                  console.log('查看');
-                }}
-              />
-              <img
-                src="https://s.xinc818.com/files/webcilklz19gz7rldus/复制_o.svg"
-                alt="复制"
-                onClick={() => {
-                  console.log('复制');
-                }}
-              />
-            </div>
-          </div>
-        );
-      },
-      sorter: (a, b) => a.age - b.age,
-  },
-  {
-    title: 'Address',
-    dataIndex: 'address',
-  },
-];
+interface IViewTableCellData {
+  name: string;
+  value: any;
+}
 
 const data: DataType[] = [];
 for (let i = 0; i < 100; i++) {
@@ -85,17 +25,124 @@ for (let i = 0; i < 100; i++) {
   });
 }
 
-const EditorRightResultTable: React.FC = () => (
-  <div
-    className={styles.tableBox}
-  >
-    <Table
-      columns={columns}
-      bordered={true}
-      dataSource={data}
-      pagination={false}
-    />
-    <div className={styles.statusBar}>Result：执行成功. Time Consuming：25ms</div>
-  </div>
-);
+const EditorRightResultTable: React.FC = () => {
+  const [viewTableCellData, setViewTableCellData] = useState<IViewTableCellData | null>(null);
+
+  // 关闭弹窗
+  const handleCancel = () => {
+    setViewTableCellData(null);
+  };
+  //复制
+  const copyTableCell = (data: IViewTableCellData) => {
+    navigator.clipboard.writeText(data?.value || viewTableCellData?.value);
+    message.success('复制成功');
+  };
+  //点击唤起modal
+  const viewTableCell = (data: IViewTableCellData) => {
+    setViewTableCellData(data);
+  };
+
+  const columns: ColumnsType<DataType> = [
+    {
+      title: 'Name',
+      dataIndex: 'name',
+      width: 130,
+      render: (value: any, row: any, rowIndex: number) => {
+        return (
+          <div className={styles.tableItem}>
+            <div>{value}</div>
+            <div className={styles.tableHoverBox}>
+              <img
+                src="https://s.xinc818.com/files/webcilklz16y4pxm3zv/位图 (1).svg"
+                alt="查看"
+                onClick={() => {
+                  viewTableCell({ name: value, value });
+                }}
+              />
+              <img
+                src="https://s.xinc818.com/files/webcilklz19gz7rldus/复制_o.svg"
+                alt="复制"
+                onClick={() => {
+                    copyTableCell({ name: value, value })
+                }}
+              />
+            </div>
+          </div>
+        );
+      },
+      sorter: (a, b) => a.name.length - b.name.length,
+    },
+    {
+      title: 'Age',
+      dataIndex: 'age',
+      width: 130,
+      render: (value: any, row: any, rowIndex: number) => {
+        return (
+          <div className={styles.tableItem}>
+            <div>{value}</div>
+            <div className={styles.tableHoverBox}>
+              <img
+                src="https://s.xinc818.com/files/webcilklz16y4pxm3zv/位图 (1).svg"
+                alt="查看"
+                onClick={() => {
+                  viewTableCell({ name: value, value });
+                }}
+              />
+              <img
+                src="https://s.xinc818.com/files/webcilklz19gz7rldus/复制_o.svg"
+                alt="复制"
+                onClick={() => {
+                    copyTableCell({ name: value, value })
+                }}
+              />
+            </div>
+          </div>
+        );
+      },
+      sorter: (a, b) => a.age - b.age,
+    },
+    {
+      title: 'Address',
+      dataIndex: 'address',
+    },
+  ];
+  return (
+    <div className={styles.tableBox}>
+      <Table columns={columns} bordered={true} dataSource={data} pagination={false} />
+      <div className={styles.statusBar}>Result：执行成功. Time Consuming：25ms</div>
+      <Modal
+        title={viewTableCellData?.name}
+        open={!!viewTableCellData?.name}
+        onCancel={handleCancel}
+        width="60vw"
+        maskClosable={false}
+        footer={
+          <>
+            {
+              <Button
+                onClick={copyTableCell.bind(null, viewTableCellData!)}
+                className={styles.cancel}
+              >
+                Copy
+              </Button>
+            }
+          </>
+        }
+      >
+        <div className={styles.monacoEditor}>
+          {/* <MonacoEditor
+            id="view_table-Cell_data"
+            appendValue={{
+              text: viewTableCellData?.value,
+              range: 'reset',
+            }}
+            options={{
+              readOnly: true,
+            }}
+          /> */}
+        </div>
+      </Modal>
+    </div>
+  );
+};
 export default EditorRightResultTable;
